change save snapshot default behavior
The default behavior is now to supress saving a galaxy's
timeseries data (previous behavior was to save automatically).
Timeseries data will be saved if the user passes
`--save-snapshots` at the command line.

To achieve this,
1. the command-line option `--no-snapshots` changes
to `--save-snapshots`, which is stored as `True` when passed by
the user.
2. the `if not (opts.no_snapshots):` logic changes to
`if opts.save_snapshot`
3. remove the dependence on the `--verbose` option